Implement terraform plugin dir
In default usage, terraform init downloads and installs the plugins for
any providers used in the configuration automatically. In automation
environments, it can be desirable to disable this behavior and instead
provide a fixed set of plugins already installed on the system where
Terraform is running. This then avoids the overhead of re-downloading
the plugins on each execution, and allows the system administrator to
control which plugins are available.

Additionally, this helps to speed up testing by avoiding unnecessary
downloading of terraform plugins for every test.

Authored-by: Kris Macoskey <kmacoskey@pivotal.io>